# fooddrop-backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/health": {
      "get": {
        "operationId": "HealthController_getHealth",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponseDto"
                }
              }
            }
          },
          "503": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "health"
        ]
      }
    },
    "/ingredients": {
      "get": {
        "operationId": "IngredientsController_search",
        "parameters": [
          {
            "name": "q",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "maxLength": 100
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 20,
              "type": "integer",
              "minimum": 1,
              "maximum": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Ingredient"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "ingredients"
        ]
      },
      "post": {
        "operationId": "IngredientsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateIngredient"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ingredient"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "ingredients"
        ]
      }
    },
    "/tags": {
      "get": {
        "operationId": "TagsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TagDimension"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "tags"
        ]
      }
    },
    "/recipes": {
      "get": {
        "operationId": "RecipesController_list",
        "parameters": [
          {
            "name": "tags",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "rarity",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "maxMinutes",
            "required": false,
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "exclusiveMinimum": true,
              "maximum": 10080
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "maxLength": 100
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "maxLength": 200
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 20,
              "type": "integer",
              "minimum": 1,
              "maximum": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeList"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "recipes"
        ]
      },
      "post": {
        "operationId": "RecipesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecipeInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "recipes"
        ]
      }
    },
    "/recipes/{id}": {
      "get": {
        "operationId": "RecipesController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "recipes"
        ]
      },
      "put": {
        "operationId": "RecipesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecipeInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "recipes"
        ]
      },
      "delete": {
        "operationId": "RecipesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "recipes"
        ]
      }
    }
  },
  "info": {
    "title": "Food Drop API",
    "description": "",
    "version": "0.1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "HealthChecksDto": {
        "type": "object",
        "properties": {
          "postgres": {
            "type": "string",
            "enum": [
              "up",
              "down"
            ]
          },
          "redis": {
            "type": "string",
            "enum": [
              "up",
              "down"
            ]
          }
        },
        "required": [
          "postgres",
          "redis"
        ]
      },
      "HealthResponseDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ok",
              "error"
            ]
          },
          "checks": {
            "$ref": "#/components/schemas/HealthChecksDto"
          }
        },
        "required": [
          "status",
          "checks"
        ]
      },
      "Ingredient": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "name": {
            "type": "string"
          },
          "aliases": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "aisle": {
            "type": "string",
            "enum": [
              "produce",
              "meat",
              "seafood",
              "dairy",
              "pantry",
              "spices",
              "frozen",
              "other"
            ]
          },
          "defaultUnit": {
            "type": "string",
            "enum": [
              "g",
              "ml",
              "piece"
            ]
          },
          "densityGPerMl": {
            "nullable": true,
            "type": "number"
          },
          "isFermented": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "name",
          "aliases",
          "aisle",
          "defaultUnit",
          "densityGPerMl",
          "isFermented"
        ],
        "additionalProperties": false
      },
      "SearchIngredientsQuery": {
        "type": "object",
        "properties": {
          "q": {
            "type": "string",
            "maxLength": 100
          },
          "limit": {
            "default": 20,
            "type": "integer",
            "minimum": 1,
            "maximum": 50
          }
        }
      },
      "CreateIngredient": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "aliases": {
            "default": [],
            "maxItems": 20,
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100
            }
          },
          "aisle": {
            "default": "other",
            "type": "string",
            "enum": [
              "produce",
              "meat",
              "seafood",
              "dairy",
              "pantry",
              "spices",
              "frozen",
              "other"
            ]
          },
          "defaultUnit": {
            "default": "g",
            "type": "string",
            "enum": [
              "g",
              "ml",
              "piece"
            ]
          },
          "densityGPerMl": {
            "default": null,
            "nullable": true,
            "type": "number",
            "minimum": 0,
            "exclusiveMinimum": true,
            "maximum": 25
          }
        },
        "required": [
          "name"
        ]
      },
      "TagDimension": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "slug": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "integer",
                  "minimum": -9007199254740991,
                  "maximum": 9007199254740991
                },
                "slug": {
                  "type": "string"
                },
                "label": {
                  "type": "string"
                }
              },
              "required": [
                "id",
                "slug",
                "label"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "id",
          "slug",
          "label",
          "tags"
        ],
        "additionalProperties": false
      },
      "RecipeInput": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "description": {
            "nullable": true,
            "type": "string",
            "maxLength": 2000
          },
          "imageUrl": {
            "nullable": true,
            "type": "string",
            "maxLength": 2000,
            "format": "uri"
          },
          "sourceUrl": {
            "nullable": true,
            "type": "string",
            "maxLength": 2000,
            "format": "uri"
          },
          "baseServings": {
            "default": 2,
            "type": "integer",
            "minimum": 1,
            "maximum": 100
          },
          "totalMinutes": {
            "type": "integer",
            "minimum": 1,
            "maximum": 10080
          },
          "difficulty": {
            "type": "integer",
            "minimum": 1,
            "maximum": 5
          },
          "steps": {
            "minItems": 1,
            "maxItems": 100,
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "text": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 2000
                },
                "timerSeconds": {
                  "type": "integer",
                  "minimum": 1,
                  "maximum": 86400
                },
                "timerLabel": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 100
                }
              },
              "required": [
                "text"
              ]
            }
          },
          "ingredients": {
            "default": [],
            "maxItems": 100,
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "ingredientId": {
                  "type": "string",
                  "format": "uuid",
                  "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                },
                "quantity": {
                  "anyOf": [
                    {
                      "type": "number",
                      "minimum": 0,
                      "maximum": 1000000
                    },
                    {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 20
                    }
                  ]
                },
                "unit": {
                  "nullable": true,
                  "type": "string",
                  "maxLength": 30
                },
                "note": {
                  "nullable": true,
                  "type": "string",
                  "maxLength": 200
                }
              },
              "required": [
                "ingredientId",
                "quantity"
              ]
            }
          },
          "tagIds": {
            "default": [],
            "maxItems": 30,
            "type": "array",
            "items": {
              "type": "integer",
              "minimum": 0,
              "exclusiveMinimum": true,
              "maximum": 9007199254740991
            }
          }
        },
        "required": [
          "title",
          "totalMinutes",
          "difficulty",
          "steps"
        ]
      },
      "ListRecipesQuery": {
        "type": "object",
        "properties": {
          "tags": {
            "type": "string"
          },
          "rarity": {
            "type": "string"
          },
          "maxMinutes": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": true,
            "maximum": 10080
          },
          "q": {
            "type": "string",
            "maxLength": 100
          },
          "cursor": {
            "type": "string",
            "maxLength": 200
          },
          "limit": {
            "default": 20,
            "type": "integer",
            "minimum": 1,
            "maximum": 50
          }
        }
      },
      "RecipeDetail": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "nullable": true,
            "type": "string"
          },
          "imageUrl": {
            "nullable": true,
            "type": "string"
          },
          "baseServings": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "totalMinutes": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "difficulty": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "rarity": {
            "type": "string",
            "enum": [
              "white",
              "blue",
              "purple",
              "pink",
              "red"
            ]
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "integer",
                  "minimum": -9007199254740991,
                  "maximum": 9007199254740991
                },
                "slug": {
                  "type": "string"
                },
                "label": {
                  "type": "string"
                },
                "dimension": {
                  "type": "string"
                }
              },
              "required": [
                "id",
                "slug",
                "label",
                "dimension"
              ],
              "additionalProperties": false
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
          },
          "sourceUrl": {
            "nullable": true,
            "type": "string"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
          },
          "steps": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "order": {
                  "type": "integer",
                  "minimum": -9007199254740991,
                  "maximum": 9007199254740991
                },
                "text": {
                  "type": "string"
                },
                "timerSeconds": {
                  "type": "integer",
                  "minimum": -9007199254740991,
                  "maximum": 9007199254740991
                },
                "timerLabel": {
                  "type": "string"
                }
              },
              "required": [
                "order",
                "text"
              ],
              "additionalProperties": false
            }
          },
          "ingredients": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "ingredient": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    "name": {
                      "type": "string"
                    },
                    "aisle": {
                      "type": "string",
                      "enum": [
                        "produce",
                        "meat",
                        "seafood",
                        "dairy",
                        "pantry",
                        "spices",
                        "frozen",
                        "other"
                      ]
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "aisle"
                  ],
                  "additionalProperties": false
                },
                "quantity": {
                  "type": "number"
                },
                "unit": {
                  "type": "string",
                  "enum": [
                    "g",
                    "ml",
                    "piece"
                  ]
                },
                "note": {
                  "nullable": true,
                  "type": "string"
                }
              },
              "required": [
                "ingredient",
                "quantity",
                "unit",
                "note"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "id",
          "title",
          "description",
          "imageUrl",
          "baseServings",
          "totalMinutes",
          "difficulty",
          "rarity",
          "tags",
          "createdAt",
          "sourceUrl",
          "updatedAt",
          "steps",
          "ingredients"
        ],
        "additionalProperties": false
      },
      "RecipeList": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid",
                  "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                },
                "title": {
                  "type": "string"
                },
                "description": {
                  "nullable": true,
                  "type": "string"
                },
                "imageUrl": {
                  "nullable": true,
                  "type": "string"
                },
                "baseServings": {
                  "type": "integer",
                  "minimum": -9007199254740991,
                  "maximum": 9007199254740991
                },
                "totalMinutes": {
                  "type": "integer",
                  "minimum": -9007199254740991,
                  "maximum": 9007199254740991
                },
                "difficulty": {
                  "type": "integer",
                  "minimum": -9007199254740991,
                  "maximum": 9007199254740991
                },
                "rarity": {
                  "type": "string",
                  "enum": [
                    "white",
                    "blue",
                    "purple",
                    "pink",
                    "red"
                  ]
                },
                "tags": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer",
                        "minimum": -9007199254740991,
                        "maximum": 9007199254740991
                      },
                      "slug": {
                        "type": "string"
                      },
                      "label": {
                        "type": "string"
                      },
                      "dimension": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "id",
                      "slug",
                      "label",
                      "dimension"
                    ],
                    "additionalProperties": false
                  }
                },
                "createdAt": {
                  "type": "string",
                  "format": "date-time",
                  "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
                }
              },
              "required": [
                "id",
                "title",
                "description",
                "imageUrl",
                "baseServings",
                "totalMinutes",
                "difficulty",
                "rarity",
                "tags",
                "createdAt"
              ],
              "additionalProperties": false
            }
          },
          "nextCursor": {
            "nullable": true,
            "type": "string"
          }
        },
        "required": [
          "items",
          "nextCursor"
        ],
        "additionalProperties": false
      }
    }
  }
}